Commonize more of SearchAutoComplete
