+++ packages/lesswrong/components/search/UsersSearchAutoComplete.jsx
@@ -1,27 +1,14 @@
 import React from 'react';
-import { registerComponent, Components, getSetting } from 'meteor/vulcan:core'
-import { InstantSearch } from 'react-instantsearch-dom';
+import { registerComponent, Components } from 'meteor/vulcan:core'
 import { algoliaIndexNames } from '../../lib/algoliaIndexNames.js';
 
 const UsersSearchAutoComplete = ({clickAction, label}) => {
-  const algoliaAppId = getSetting('algolia.appId')
-  const algoliaSearchKey = getSetting('algolia.searchKey')
-
-  if(!algoliaAppId) {
-    return <div>User search is disabled (Algolia App ID not configured on server)</div>
-  }
-
-  return <InstantSearch
+  return <Components.SearchAutoComplete
     indexName={algoliaIndexNames.Users}
-    appId={algoliaAppId}
-    apiKey={algoliaSearchKey}
-  >
-    <Components.SearchAutoComplete
-      clickAction={suggestion => clickAction(suggestion.objectID)}
-      renderSuggestion={hit => <Components.UsersAutoCompleteHit document={hit} />}
-      placeholder={label || "Search for Users"}
-    />
-  </InstantSearch>
+    clickAction={suggestion => clickAction(suggestion.objectID)}
+    renderSuggestion={hit => <Components.UsersAutoCompleteHit document={hit} />}
+    placeholder={label || "Search for Users"}
+  />
 }
 
 registerComponent("UsersSearchAutoComplete", UsersSearchAutoComplete);
